[android] News package is renamed to onboarding
