+++ android/src/com/mapswithme/maps/MwmActivity.java
@@ -72,8 +72,8 @@ import com.mapswithme.maps.maplayer.traffic.OnTrafficLayerToggleListener;
 import com.mapswithme.maps.maplayer.traffic.TrafficManager;
 import com.mapswithme.maps.maplayer.traffic.widget.TrafficButton;
 import com.mapswithme.maps.metrics.UserActionsLogger;
-import com.mapswithme.maps.news.IntroductionDialogFragment;
-import com.mapswithme.maps.news.IntroductionScreenFactory;
+import com.mapswithme.maps.onboarding.IntroductionDialogFragment;
+import com.mapswithme.maps.onboarding.IntroductionScreenFactory;
 import com.mapswithme.maps.promo.Promo;
 import com.mapswithme.maps.promo.PromoAfterBooking;
 import com.mapswithme.maps.promo.PromoBookingDialogFragment;
