+++ android/src/com/mapswithme/maps/SplashActivity.java
@@ -22,9 +22,9 @@ import com.mapswithme.maps.base.Detachable;
 import com.mapswithme.maps.downloader.UpdaterDialogFragment;
 import com.mapswithme.maps.editor.ViralFragment;
 import com.mapswithme.maps.location.LocationHelper;
-import com.mapswithme.maps.news.BaseNewsFragment;
-import com.mapswithme.maps.news.NewsFragment;
-import com.mapswithme.maps.news.WelcomeDialogFragment;
+import com.mapswithme.maps.onboarding.BaseNewsFragment;
+import com.mapswithme.maps.onboarding.NewsFragment;
+import com.mapswithme.maps.onboarding.WelcomeDialogFragment;
 import com.mapswithme.maps.permissions.PermissionsDialogFragment;
 import com.mapswithme.maps.permissions.StoragePermissionsDialogFragment;
 import com.mapswithme.util.Config;
