+++ android/src/com/mapswithme/maps/downloader/UpdaterDialogFragment.java
@@ -17,7 +17,7 @@ import android.widget.TextView;
 import com.mapswithme.maps.Framework;
 import com.mapswithme.maps.R;
 import com.mapswithme.maps.base.BaseMwmDialogFragment;
-import com.mapswithme.maps.news.BaseNewsFragment;
+import com.mapswithme.maps.onboarding.BaseNewsFragment;
 import com.mapswithme.maps.widget.WheelProgressView;
 import com.mapswithme.util.Constants;
 import com.mapswithme.util.StringUtils;
