+++ android/src/com/mapswithme/maps/onboarding/PromoButton.java
@@ -1,4 +1,4 @@
-package com.mapswithme.maps.news;
+package com.mapswithme.maps.onboarding;
 
 import androidx.annotation.NonNull;
